Error de casting corregido

+++ GameDesignPattern/src/base/Enemy.java
@@ -18,11 +18,11 @@ public class Enemy extends Character{
 	public Stats levelStats(Stats stats, World world) {
 		Stats modifiedStats = new Stats();
 		
-		modifiedStats.setMaxLife(stats.getMaxLife() * world.getComplexFactor());
-		modifiedStats.setLife(stats.getLife() * world.getComplexFactor());
-		modifiedStats.setAttack(stats.getAttack() * world.getComplexFactor());
-		modifiedStats.setDefense(stats.getDefense() * world.getComplexFactor());
-		modifiedStats.setSpeed(stats.getSpeed() * world.getComplexFactor());
+		modifiedStats.setMaxLife((int)(stats.getMaxLife() * world.getComplexFactor()));
+		modifiedStats.setLife((int)(stats.getLife() * world.getComplexFactor()));
+		modifiedStats.setAttack((int)(stats.getAttack() * world.getComplexFactor()));
+		modifiedStats.setDefense((int)(stats.getDefense() * world.getComplexFactor()));
+		modifiedStats.setSpeed((int)(stats.getSpeed() * world.getComplexFactor()));
 		
 		return modifiedStats;
 	}
